fix: Headings should reflect content correctly when viewing old document revision
