+++ app/models/Revision.js
@@ -1,4 +1,6 @@
 // @flow
+import { computed } from 'mobx';
+import getHeadingsForText from 'shared/utils/getHeadingsForText';
 import BaseModel from './BaseModel';
 import User from './User';
 
@@ -9,6 +11,11 @@ class Revision extends BaseModel {
   text: string;
   createdAt: string;
   createdBy: User;
+
+  @computed
+  get headings() {
+    return getHeadingsForText(this.text);
+  }
 }
 
 export default Revision;
